requirements.txt
transformers
accelerate
torch
loguru
requests
python-dotenv
